Reformat BaseWorker test
